멤버탈퇴, 관리자화면

+++ src/main/java/what/the/bus/member/controller/DeleteMemberController.java
@@ -1,21 +1,31 @@
 package what.the.bus.member.controller;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
 import what.the.bus.member.MemberVO;
 import what.the.bus.member.service.DeleteMemberService;
 
 @Controller
+@SessionAttributes("member")
 public class DeleteMemberController {
 	@Autowired
 	private DeleteMemberService memberService;
 
 	// 회원탈퇴
 	@RequestMapping("/view/**/deleteMember.do")
-	public String deleteMember(MemberVO vo) {
-		memberService.deleteMember(vo);
-		return "redirect:../main/main.jsp";
+	public String deleteMember(@ModelAttribute("member") MemberVO vo, HttpSession session) {
+		if(memberService.pwCheckMember(vo) == true) {
+			session.invalidate();	
+			memberService.deleteMember(vo);
+			return "redirect:../main/main.jsp";
+	} else {
+			return "redirect:../member/errorPW.jsp";
 	}
+}
 }
